修改文档规范，补充免责声明，修改统计值数据类型

+++ include/kdadk_feature_extract.h
@@ -47,7 +47,7 @@ typedef struct {
 } rawbow_list;
 
 typedef struct {
-    uint32_t total_packets;   /* 总包数 */
+    uint64_t total_packets;   /* 总包数 */
     uint64_t total_bytes;     /* 总字节数 */
     uint32_t decoded_packets; /* 成功解析的包数 */
     uint32_t extracted_flows; /* 已提取的流数 */
